+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes by Version
 -------------------
 
 - Added Thrift support to ``tcurl.py`` and re-worked the script's arguments.
+- Changed minimum required version of Tornado to 4.2.
 
 
 0.17.3 (unreleased)